- CurrentSong class added. - Current song length is now stored as an Int.

+++ app/src/main/java/com/minichain/minichainsplayer/SongFile.kt
@@ -4,9 +4,9 @@ class SongFile {
     var path: String
     var songName: String
     var format: String
-    var length: Long = 0
+    var length: Int = 0
 
-    constructor(path: String, songName: String, format: String, length: Long) {
+    constructor(path: String, songName: String, format: String, length: Int) {
         this.path = path
         this.songName = songName
         this.format = format
